add lib/trajectory/plotting.ts with generate_plot_series, should_hide_plot functions
- move `get_label_with_unit` function to lib/labels.ts
- move `plot_colors` constant to lib/colors/index.ts
- simplify `Trajectory.svelte` by importing new plotting utils
